Fix guards.ts + added descriptions in locale yaml indexes
