+++ ts/utils/guards.ts
@@ -1,24 +1,6 @@
-import { Wallet } from "../types/pagopa";
-
 /**
  * A type guard that filters out undefined and null from a type T
  */
 export function isDefined<T, O extends NonNullable<T>>(v: T): v is O {
   return v !== null && v !== undefined;
 }
-
-export const isExpiredCard = (w: Wallet) => {
-  const today: Date = new Date();
-  const cmpM: number = today.getMonth() + 1;
-  const cmpY: number = parseInt(
-    today
-      .getFullYear()
-      .toString()
-      .slice(2),
-    10
-  );
-  return (
-    parseInt(w.creditCard.expireMonth, 10) > cmpM &&
-    parseInt(w.creditCard.expireYear, 10) <= cmpY
-  );
-};
